Fixed bug

+++ src/main/scala/jp/ijufumi/openreports/config/Config.scala
@@ -12,7 +12,7 @@ object Config {
   val API_TOKEN_EXPIRATION_SEC: Integer =
     Integer.parseInt(sys.env.getOrElse("API_TOKEN_EXPIRATION_SEC", "3600"))
   // for cache
-  val CACHE_TTL_SEC: Integer = Integer.parseInt(sys.env.getOrElse("CACHE_TTL_SEC", "3600"))
+  val CACHE_TTL_SEC: Long = Integer.parseInt(sys.env.getOrElse("CACHE_TTL_SEC", "3600"))
   // for api
   val API_TOKEN_KEY = "x-api-token"
   // for google auth
